Extract TransactionCard component
